Implement opening notification modal if the user opens it by tapping a notification
